+++ src/components/login/Login.jsx
@@ -1,120 +1,99 @@
-// import React, { useState } from "react";
+import React, { useState } from "react";
+import { createPortal } from "react-dom";
+import { Link } from "react-router-dom";
 
+const modalElement = document.getElementById("modalLogin");
 
-// export const Login = ({setShow}) => {
+export const Login = ({ show, setShow }) => {
+  const closeModal = () => {
+    setShow(false);
+  };
 
+  const handleFormSubmit = (e) => {
+    e.preventDefault();
+    localStorage.setItem("userData", JSON.stringify(user));
+    closeModal();
+  };
+  
+  
+  
+  const [user, setUser] = useState({
+    email: "",
+    password: "",
+  });
 
- 
+  const handleInputChange = (e) => {
+    const { name, value } = e.target;
+    setUser((prevState) => ({
+      ...prevState,
+      [name]: value,
+    }));
+    console.log(user)
+  };
+  
 
-//   const closeModal = () => {
-//     setShow(false);
-//   };
+  return (
+    <div className="relative">
+      {show &&
+        createPortal(
+          <div className="flex w-[96%] h-[30rem] sm:w-[30rem] items-center justify-around text-white rounded-md fixed top-[6.4rem] sm:top-[10rem] flex-col right-0 left-0 sm:right-[7rem] sm:left-[7rem] mx-auto sm:mx-auto backdrop-opacity-90 bg-black/75 backdrop-blur-md">
+            <button className="bg-arena-green-400 ms-auto me-3 mt-3 sm:mt-5 rounded-sm" onClick={closeModal}>X</button>
+            <section className="flex w-full flex-col space-y-10 p-5">
+            <div className="text-center text-4xl font-medium">
+              Iniciar Sesión
+            </div>
+              <form className="space-y-10 flex flex-col"  onSubmit={handleFormSubmit}>
+                <div className="mb-4 flex flex-col gap-6">
+                  <div className="w-full transform border-b-2 bg-transparent text-lg duration-300 focus-within:border-arena-green-400">
+                    <input
+                      type="email"
+                      placeholder="Email"
+                      name="email"
+                      value={setUser.email}
+                      onChange={handleInputChange}
+                      className="w-full border-none bg-transparent outline-none focus:outline-none"
+                    />
+                  </div>
 
-//   return (
-    
-//       <div className="flex min-h-screen items-center justify-center absolute top-10 left-0 right-0 mx-auto">
-//         <div className="relative flex flex-col rounded-xl bg-transparent bg-clip-border text-gray-700 shadow-none">
-//           <h4 className="block font-sans text-2xl font-semibold leading-snug tracking-normal text-blue-gray-900 antialiased">
-//             Sign Up
-//           </h4>
-//           <button onClick={closeModal}>CLOSE</button>
-//           <p className="mt-1 block font-sans text-base font-normal leading-relaxed text-gray-700 antialiased">
-//             Enter your details to register.
-//           </p>
-//           <form className="mt-8 mb-2 w-80 max-w-screen-lg sm:w-96">
-//             <div className="mb-4 flex flex-col gap-6">
-//               <div className="relative h-11 w-full min-w-[200px]">
-//                 <input
-//                   className="peer h-full w-full rounded-md border border-blue-gray-200 bg-transparent px-3 py-3 font-sans text-sm font-normal text-blue-gray-700 outline outline-0 transition-all placeholder-shown:border placeholder-shown:border-blue-gray-200 placeholder-shown:border-t-blue-gray-200 focus:border-2 focus:border-pink-500 focus:border-t-transparent focus:outline-0 disabled:border-0 disabled:bg-blue-gray-50"
-//                   placeHolder=" "
-//                 />
-//                 <label className="before:content[' '] after:content[' '] pointer-events-none absolute left-0 -top-1.5 flex h-full w-full select-none text-[11px] font-normal leading-tight text-blue-gray-400 transition-all before:pointer-events-none before:mt-[6.5px] before:mr-1 before:box-border before:block before:h-1.5 before:w-2.5 before:rounded-tl-md before:border-t before:border-l before:border-blue-gray-200 before:transition-all after:pointer-events-none after:mt-[6.5px] after:ml-1 after:box-border after:block after:h-1.5 after:w-2.5 after:flex-grow after:rounded-tr-md after:border-t after:border-r after:border-blue-gray-200 after:transition-all peer-placeholder-shown:text-sm peer-placeholder-shown:leading-[4.1] peer-placeholder-shown:text-blue-gray-500 peer-placeholder-shown:before:border-transparent peer-placeholder-shown:after:border-transparent peer-focus:text-[11px] peer-focus:leading-tight peer-focus:text-pink-500 peer-focus:before:border-t-2 peer-focus:before:border-l-2 peer-focus:before:!border-pink-500 peer-focus:after:border-t-2 peer-focus:after:border-r-2 peer-focus:after:!border-pink-500 peer-disabled:text-transparent peer-disabled:before:border-transparent peer-disabled:after:border-transparent peer-disabled:peer-placeholder-shown:text-blue-gray-500">
-//                   Name
-//                 </label>
-//               </div>
-//               <div className="relative h-11 w-full min-w-[200px]">
-//                 <input
-//                   className="peer h-full w-full rounded-md border border-blue-gray-200 bg-transparent px-3 py-3 font-sans text-sm font-normal text-blue-gray-700 outline outline-0 transition-all placeholder-shown:border placeholder-shown:border-blue-gray-200 placeholder-shown:border-t-blue-gray-200 focus:border-2 focus:border-pink-500 focus:border-t-transparent focus:outline-0 disabled:border-0 disabled:bg-blue-gray-50"
-//                   placeHolder=" "
-//                 />
-//                 <label className="before:content[' '] after:content[' '] pointer-events-none absolute left-0 -top-1.5 flex h-full w-full select-none text-[11px] font-normal leading-tight text-blue-gray-400 transition-all before:pointer-events-none before:mt-[6.5px] before:mr-1 before:box-border before:block before:h-1.5 before:w-2.5 before:rounded-tl-md before:border-t before:border-l before:border-blue-gray-200 before:transition-all after:pointer-events-none after:mt-[6.5px] after:ml-1 after:box-border after:block after:h-1.5 after:w-2.5 after:flex-grow after:rounded-tr-md after:border-t after:border-r after:border-blue-gray-200 after:transition-all peer-placeholder-shown:text-sm peer-placeholder-shown:leading-[4.1] peer-placeholder-shown:text-blue-gray-500 peer-placeholder-shown:before:border-transparent peer-placeholder-shown:after:border-transparent peer-focus:text-[11px] peer-focus:leading-tight peer-focus:text-pink-500 peer-focus:before:border-t-2 peer-focus:before:border-l-2 peer-focus:before:!border-pink-500 peer-focus:after:border-t-2 peer-focus:after:border-r-2 peer-focus:after:!border-pink-500 peer-disabled:text-transparent peer-disabled:before:border-transparent peer-disabled:after:border-transparent peer-disabled:peer-placeholder-shown:text-blue-gray-500">
-//                   Email
-//                 </label>
-//               </div>
-//               <div className="relative h-11 w-full min-w-[200px]">
-//                 <input
-//                   type="password"
-//                   className="peer h-full w-full rounded-md border border-blue-gray-200 bg-transparent px-3 py-3 font-sans text-sm font-normal text-blue-gray-700 outline outline-0 transition-all placeholder-shown:border placeholder-shown:border-blue-gray-200 placeholder-shown:border-t-blue-gray-200 focus:border-2 focus:border-pink-500 focus:border-t-transparent focus:outline-0 disabled:border-0 disabled:bg-blue-gray-50"
-//                   placeHolder=" "
-//                 />
-//                 <label className="before:content[' '] after:content[' '] pointer-events-none absolute left-0 -top-1.5 flex h-full w-full select-none text-[11px] font-normal leading-tight text-blue-gray-400 transition-all before:pointer-events-none before:mt-[6.5px] before:mr-1 before:box-border before:block before:h-1.5 before:w-2.5 before:rounded-tl-md before:border-t before:border-l before:border-blue-gray-200 before:transition-all after:pointer-events-none after:mt-[6.5px] after:ml-1 after:box-border after:block after:h-1.5 after:w-2.5 after:flex-grow after:rounded-tr-md after:border-t after:border-r after:border-blue-gray-200 after:transition-all peer-placeholder-shown:text-sm peer-placeholder-shown:leading-[4.1] peer-placeholder-shown:text-blue-gray-500 peer-placeholder-shown:before:border-transparent peer-placeholder-shown:after:border-transparent peer-focus:text-[11px] peer-focus:leading-tight peer-focus:text-pink-500 peer-focus:before:border-t-2 peer-focus:before:border-l-2 peer-focus:before:!border-pink-500 peer-focus:after:border-t-2 peer-focus:after:border-r-2 peer-focus:after:!border-pink-500 peer-disabled:text-transparent peer-disabled:before:border-transparent peer-disabled:after:border-transparent peer-disabled:peer-placeholder-shown:text-blue-gray-500">
-//                   Password
-//                 </label>
-//               </div>
-//             </div>
-//             <div className="inline-flex items-center">
-//               <label
-//                 className="relative -ml-2.5 flex cursor-pointer items-center rounded-full p-3"
-//                 for="checkbox"
-//                 data-ripple-dark="true"
-//               >
-//                 <input
-//                   type="checkbox"
-//                   className="before:content[''] peer relative h-5 w-5 cursor-pointer appearance-none rounded-md border border-blue-gray-200 transition-all before:absolute before:top-2/4 before:left-2/4 before:block before:h-12 before:w-12 before:-translate-y-2/4 before:-translate-x-2/4 before:rounded-full before:bg-blue-gray-500 before:opacity-0 before:transition-opacity checked:border-pink-500 checked:bg-pink-500 checked:before:bg-pink-500 hover:before:opacity-10"
-//                   id="checkbox"
-//                 />
-//                 <span className="pointer-events-none absolute top-2/4 left-2/4 -translate-y-2/4 -translate-x-2/4 text-white opacity-0 transition-opacity peer-checked:opacity-100">
-//                   <svg
-//                     xmlns="http://www.w3.org/2000/svg"
-//                     className="h-3.5 w-3.5"
-//                     viewBox="0 0 20 20"
-//                     fill="currentColor"
-//                     stroke="currentColor"
-//                     stroke-width="1"
-//                   >
-//                     <path
-//                       fill-rule="evenodd"
-//                       d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z"
-//                       clip-rule="evenodd"
-//                     ></path>
-//                   </svg>
-//                 </span>
-//               </label>
-//               <label
-//                 className="mt-px cursor-pointer select-none font-light text-gray-700"
-//                 for="checkbox"
-//               >
-//                 <p className="flex items-center font-sans text-sm font-normal leading-normal text-gray-700 antialiased">
-//                   I agree the
-//                   <a
-//                     className="font-semibold transition-colors hover:text-pink-500"
-//                     href="#"
-//                   >
-//                     &nbsp;Terms and Conditions
-//                   </a>
-//                 </p>
-//               </label>
-//             </div>
-//             <button
-//               className="mt-6 block w-full select-none rounded-lg bg-pink-500 py-3 px-6 text-center align-middle font-sans text-xs font-bold uppercase text-white shadow-md shadow-pink-500/20 transition-all hover:shadow-lg hover:shadow-pink-500/40 focus:opacity-[0.85] focus:shadow-none active:opacity-[0.85] active:shadow-none disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
-//               type="button"
-//               data-ripple-light="true"
-//             >
-//               Register
-//             </button>
-//             <p className="mt-4 block text-center font-sans text-base font-normal leading-relaxed text-gray-700 antialiased">
-//               Already have an account?
-//               <a
-//                 className="font-semibold text-pink-500 transition-colors hover:text-blue-700"
-//                 href="#"
-//               >
-//                 Sign In
-//               </a>
-//             </p>
-//           </form>
-//         </div>
-//       </div>
-      
-    
-//   );
-// };
+                  <div className="w-full transform border-b-2 bg-transparent text-lg duration-300 focus-within:border-arena-green-400">
+                    <input
+                      type="password"
+                      name="password"
+                      placeholder="Password"
+                      value={setUser.password}
+                      onChange={handleInputChange}
+                      className="w-full border-none bg-transparent outline-none focus:outline-none"
+                    />
+                  </div>
+                </div>
+                <div className="inline-flex items-center"></div>
+                <button
+                  className="transform rounded-sm bg-arena-green-400 py-2 font-bold duration-300 hover:bg-arena-green-600 "
+                  type="submit"
+                >
+                  Iniciar Sesion
+                </button>
+              </form>
+            </section>
+              <Link
+                to="#"
+                className="transform text-center font-semibold text-gray-500 duration-300 hover:text-gray-300"
+              >
+                OLVIDASTE TU PASSWORD?
+              </Link>
+              <p className="text-center text-lg">
+              Aún no tienes cuenta?
+              <Link
+                to="/registro"
+                className="font-medium ms-2 text-arena-green-300 underline-offset-4 hover:underline"
+              >
+                Regístrate
+              </Link>
+            </p>
+            
+          </div>,
+          modalElement
+        )}
+    </div>
+  );
+};
